Move host search methods into FiberTreeReflection
I'm going to need getHostSibling.
